millisecond timestamps
With a precision of only seconds, indexers might not be able to properly
order the logs when using the ConsoleLogFilter.
Use by default timestamps with milliseconds precision for new installations of the plugin.
When migrating from old ToolInstallation way of configuration,
millisecond precision is turned of to avoid possible problems with
existing setups.