[JENKINS-51793] explicit enable
Explicitly enable indexing
When a 1.x version of the plugin was installed but never configured and then one updates to 2.0
the configuration is initially not saveable.
Also when installing the first time one couldn't save without
configuring the plugin. With the enable option it is initially disabled
and one can save without problems.